+++ theatre_management_system/src/main/java/com/project/theatre_management_system/service/MovieService.java
@@ -8,6 +8,7 @@ import com.project.theatre_management_system.dao.MovieDao;
 import com.project.theatre_management_system.dto.Movie;
 import com.project.theatre_management_system.dto.Review;
 import com.project.theatre_management_system.dto.Viewer;
+import com.project.theatre_management_system.exception.MovieIdNotFound;
 import com.project.theatre_management_system.util.ResponseStructure;
 import com.project.theatre_management_system.util.ResponseStructureList;
 
@@ -15,10 +16,10 @@ import com.project.theatre_management_system.util.ResponseStructureList;
 public class MovieService {
 	@Autowired
 	MovieDao movieDao;
-	
+
 	@Autowired
 	ResponseStructure<Movie> responseStructure;
-	
+
 	@Autowired
 	ResponseStructureList<Movie> responseStructureList;
 
@@ -28,35 +29,35 @@ public class MovieService {
 		responseStructure.setData(movieDao.saveMovie(movie));
 		return responseStructure;
 	}
-	
+
 	public ResponseStructure<Movie> addExistingScreenToExistingMovie(int screenId, int movieId) {
 		responseStructure.setStatusCode(HttpStatus.FOUND.value());
 		responseStructure.setMessage("Successfully added Existing Screen To Existing Movie");
 		responseStructure.setData(movieDao.addExistingScreenToExistingMovie(screenId, movieId));
 		return responseStructure;
 	}
-	
+
 	public ResponseStructure<Movie> addExistingViewerToExistingMovie(int viewerId, int movieId) {
 		responseStructure.setStatusCode(HttpStatus.FOUND.value());
 		responseStructure.setMessage("Successfully added Existing Viewer To Existing Movie");
 		responseStructure.setData(movieDao.addExistingViewerToExistingMovie(viewerId, movieId));
 		return responseStructure;
 	}
-	
+
 	public ResponseStructure<Movie> addNewViewerToExistingMovie(int movieId, Viewer newViewer) {
 		responseStructure.setStatusCode(HttpStatus.FOUND.value());
 		responseStructure.setMessage("Successfully added new Viewer To Existing Movie");
 		responseStructure.setData(movieDao.addNewViewerToExistingMovie(movieId, newViewer));
 		return responseStructure;
 	}
-	
+
 	public ResponseStructure<Movie> addExistingReviewToExistingMovie(int reviewId, int movieId) {
 		responseStructure.setStatusCode(HttpStatus.FOUND.value());
 		responseStructure.setMessage("Successfully added Existing Review To Existing Movie");
 		responseStructure.setData(movieDao.addExistingReviewToExistingMovie(reviewId, movieId));
 		return responseStructure;
 	}
-	
+
 	public ResponseStructure<Movie> addNewReviewToExistingMovie(int movieId, Review newReview) {
 		responseStructure.setStatusCode(HttpStatus.FOUND.value());
 		responseStructure.setMessage("Successfully added new Review To Existing Movie");
@@ -65,10 +66,15 @@ public class MovieService {
 	}
 
 	public ResponseStructure<Movie> fetchMovieById(int movieId) {
-		responseStructure.setStatusCode(HttpStatus.FOUND.value());
-		responseStructure.setMessage("Successfullyg Movie fetched By Id");
-		responseStructure.setData(movieDao.fetchMovieById(movieId));
-		return responseStructure;
+		Movie movie = movieDao.fetchMovieById(movieId);
+		if (movie != null) {
+			responseStructure.setStatusCode(HttpStatus.FOUND.value());
+			responseStructure.setMessage("Successfullyg Movie fetched By Id");
+			responseStructure.setData(movieDao.fetchMovieById(movieId));
+			return responseStructure;
+		} else {
+			throw new MovieIdNotFound();
+		}
 	}
 
 	public ResponseStructureList<Movie> fetchAllMovie() {
@@ -79,16 +85,26 @@ public class MovieService {
 	}
 
 	public ResponseStructure<Movie> deleteMovieById(int movieId) {
-		responseStructure.setStatusCode(HttpStatus.OK.value());
-		responseStructure.setMessage("Succesfully Movie deleted by id from db");
-		responseStructure.setData(movieDao.deleteMovieById(movieId));
-		return responseStructure;
+		Movie movie = movieDao.fetchMovieById(movieId);
+		if (movie != null) {
+			responseStructure.setStatusCode(HttpStatus.OK.value());
+			responseStructure.setMessage("Succesfully Movie deleted by id from db");
+			responseStructure.setData(movieDao.deleteMovieById(movieId));
+			return responseStructure;
+		} else {
+			throw new MovieIdNotFound();
+		}
 	}
 
 	public ResponseStructure<Movie> updateMovieById(int oldMovieId, Movie newMovie) {
-		responseStructure.setStatusCode(HttpStatus.OK.value());
-		responseStructure.setMessage("Succesfully Movie updated by id in db");
-		responseStructure.setData(movieDao.updateMovieById(oldMovieId, newMovie));
-		return responseStructure;
+		Movie movie = movieDao.fetchMovieById(oldMovieId);
+		if (movie != null) {
+			responseStructure.setStatusCode(HttpStatus.OK.value());
+			responseStructure.setMessage("Succesfully Movie updated by id in db");
+			responseStructure.setData(movieDao.updateMovieById(oldMovieId, newMovie));
+			return responseStructure;
+		} else {
+			throw new MovieIdNotFound();
+		}
 	}
 }
