+++ theatre_management_system/src/main/java/com/project/theatre_management_system/service/ReviewService.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Service;
 
 import com.project.theatre_management_system.dao.ReviewDao;
 import com.project.theatre_management_system.dto.Review;
+import com.project.theatre_management_system.exception.ReviewIdNotFound;
 import com.project.theatre_management_system.util.ResponseStructure;
 import com.project.theatre_management_system.util.ResponseStructureList;
 
@@ -13,10 +14,10 @@ import com.project.theatre_management_system.util.ResponseStructureList;
 public class ReviewService {
 	@Autowired
 	ReviewDao reviewDao;
-	
+
 	@Autowired
 	ResponseStructure<Review> responseStructure;
-	
+
 	@Autowired
 	ResponseStructureList<Review> responseStructureList;
 
@@ -28,10 +29,15 @@ public class ReviewService {
 	}
 
 	public ResponseStructure<Review> fetchReviewById(int reviewId) {
-		responseStructure.setStatusCode(HttpStatus.FOUND.value());
-		responseStructure.setMessage("Succesfully Saved the Owner into db");
-		responseStructure.setData(reviewDao.fetchReviewById(reviewId));
-		return responseStructure;
+		Review review = reviewDao.fetchReviewById(reviewId);
+		if (review != null) {
+			responseStructure.setStatusCode(HttpStatus.FOUND.value());
+			responseStructure.setMessage("Succesfully Saved the Owner into db");
+			responseStructure.setData(reviewDao.fetchReviewById(reviewId));
+			return responseStructure;
+		} else {
+			throw new ReviewIdNotFound();
+		}
 	}
 
 	public ResponseStructureList<Review> fetchAllReview() {
@@ -42,16 +48,26 @@ public class ReviewService {
 	}
 
 	public ResponseStructure<Review> deleteReviewById(int reviewId) {
-		responseStructure.setStatusCode(HttpStatus.OK.value());
-		responseStructure.setMessage("Succesfully Address fetched by id from db");
-		responseStructure.setData(reviewDao.deleteRwviewById(reviewId));
-		return responseStructure;
+		Review review = reviewDao.fetchReviewById(reviewId);
+		if (review != null) {
+			responseStructure.setStatusCode(HttpStatus.OK.value());
+			responseStructure.setMessage("Succesfully Address fetched by id from db");
+			responseStructure.setData(reviewDao.deleteRwviewById(reviewId));
+			return responseStructure;
+		} else {
+			throw new ReviewIdNotFound();
+		}
 	}
 
 	public ResponseStructure<Review> updateReviewById(int oldReviewId, Review newReview) {
-		responseStructure.setStatusCode(HttpStatus.OK.value());
-		responseStructure.setMessage("Succesfully Address fetched by id from db");
-		responseStructure.setData(reviewDao.updateReviewById(oldReviewId, newReview));
-		return responseStructure;
+		Review review = reviewDao.fetchReviewById(oldReviewId);
+		if (review != null) {
+			responseStructure.setStatusCode(HttpStatus.OK.value());
+			responseStructure.setMessage("Succesfully Address fetched by id from db");
+			responseStructure.setData(reviewDao.updateReviewById(oldReviewId, newReview));
+			return responseStructure;
+		} else {
+			throw new ReviewIdNotFound();
+		}
 	}
 }
