+++ theatre_management_system/src/main/java/com/project/theatre_management_system/service/TheatreService.java
@@ -7,67 +7,83 @@ import org.springframework.stereotype.Service;
 import com.project.theatre_management_system.dao.TheatreDao;
 import com.project.theatre_management_system.dto.Branch;
 import com.project.theatre_management_system.dto.Theatre;
+import com.project.theatre_management_system.exception.TheatreIdNotFound;
 import com.project.theatre_management_system.util.ResponseStructure;
 import com.project.theatre_management_system.util.ResponseStructureList;
 
 @Service
 public class TheatreService {
-	
+
 	@Autowired
 	TheatreDao theatreDao;
-	
+
 	@Autowired
 	ResponseStructure<Theatre> responseStructure;
-	
+
 	@Autowired
 	ResponseStructureList<Theatre> responseStructureList;
-	
+
 	public ResponseStructure<Theatre> saveTheatre(Theatre theatre) {
 		responseStructure.setStatusCode(HttpStatus.CREATED.value());
 		responseStructure.setMessage("Succesfully Saved the Owner into db");
 		responseStructure.setData(theatreDao.saveTheatre(theatre));
 		return responseStructure;
 	}
-	
+
 	public ResponseStructure<Theatre> addExistingBranchToExistingTheatre(int branchId, int theatreId) {
 		responseStructure.setStatusCode(HttpStatus.FOUND.value());
 		responseStructure.setMessage("Succesfully Saved the Owner into db");
 		responseStructure.setData(theatreDao.addExistingBranchToExistingTheatre(branchId, theatreId));
 		return responseStructure;
 	}
-	
+
 	public ResponseStructure<Theatre> addNewBranchToExistingTheatre(int theatreId, Branch newBranch) {
 		responseStructure.setStatusCode(HttpStatus.FOUND.value());
 		responseStructure.setMessage("Succesfully Saved the Owner into db");
 		responseStructure.setData(theatreDao.addNewBranchToExistingTheatre(theatreId, newBranch));
 		return responseStructure;
 	}
-	
+
 	public ResponseStructure<Theatre> fetchTheatreById(int theatreId) {
-		responseStructure.setStatusCode(HttpStatus.FOUND.value());
-		responseStructure.setMessage("Succesfully Saved the Owner into db");
-		responseStructure.setData(theatreDao.fetchTheatreById(theatreId));
-		return responseStructure;
+		Theatre theatre = theatreDao.fetchTheatreById(theatreId);
+		if (theatre != null) {
+			responseStructure.setStatusCode(HttpStatus.FOUND.value());
+			responseStructure.setMessage("Succesfully Saved the Owner into db");
+			responseStructure.setData(theatreDao.fetchTheatreById(theatreId));
+			return responseStructure;
+		} else {
+			throw new TheatreIdNotFound();
+		}
 	}
-	
+
 	public ResponseStructureList<Theatre> fetchAllTheatre() {
 		responseStructureList.setStatusCode(HttpStatus.FOUND.value());
 		responseStructureList.setMessage("Succesfully Address fetched by id from db");
 		responseStructureList.setData(theatreDao.fetchAllTheatre());
 		return responseStructureList;
 	}
-	
+
 	public ResponseStructure<Theatre> deleteTheatreById(int theatreId) {
-		responseStructure.setStatusCode(HttpStatus.OK.value());
-		responseStructure.setMessage("Succesfully Address fetched by id from db");
-		responseStructure.setData(theatreDao.deleteTheatreById(theatreId));
-		return responseStructure;
+		Theatre theatre = theatreDao.fetchTheatreById(theatreId);
+		if (theatre != null) {
+			responseStructure.setStatusCode(HttpStatus.OK.value());
+			responseStructure.setMessage("Succesfully Address fetched by id from db");
+			responseStructure.setData(theatreDao.deleteTheatreById(theatreId));
+			return responseStructure;
+		} else {
+			throw new TheatreIdNotFound();
+		}
 	}
-	
+
 	public ResponseStructure<Theatre> updateTheatreById(int oldTheatreId, Theatre newTheatre) {
-		responseStructure.setStatusCode(HttpStatus.OK.value());
-		responseStructure.setMessage("Succesfully Address fetched by id from db");
-		responseStructure.setData(theatreDao.updateTheatreById(oldTheatreId, newTheatre));
-		return responseStructure;
+		Theatre theatre = theatreDao.fetchTheatreById(oldTheatreId);
+		if (theatre != null) {
+			responseStructure.setStatusCode(HttpStatus.OK.value());
+			responseStructure.setMessage("Succesfully Address fetched by id from db");
+			responseStructure.setData(theatreDao.updateTheatreById(oldTheatreId, newTheatre));
+			return responseStructure;
+		} else {
+			throw new TheatreIdNotFound();
+		}
 	}
 }
